Removed header-wrapper-2
