+++ demo/demo8.js
@@ -6,17 +6,17 @@ doc.createParagraph("First Page").pageBreak()
 doc.createParagraph("Second Page");
 
 
-var pageoneheader = new docx.Paragraph("Running head: My Title").right()
-//var tab = new docx.TextRun().tab()
+var pageoneheader = new docx.Paragraph("Running head: My Title").maxRightTabStop();
+var tab = new docx.TextRun().tab()
 var pageNumber = new docx.TextRun().pageNumber()
 
-//pageoneheader.addRun(tab);
+pageoneheader.addRun(tab);
 pageoneheader.addRun(pageNumber);
-doc.Header2.addParagraph(pageoneheader);
+doc.firstPageHeader.addParagraph(pageoneheader);
 
 var pagetwoheader = new docx.Paragraph("My Title").maxRightTabStop();
 
-//pagetwoheader.addRun(tab)
+pagetwoheader.addRun(tab)
 pagetwoheader.addRun(pageNumber)
 doc.Header.addParagraph(pagetwoheader)
 doc.Header = new docx.Paragraph("My Title")
