--- src/file/header-wrapper2.ts
@@ -1,55 +1,0 @@
-import { Header2 } from "./header2/header";
-import { IMediaData, Media } from "./media";
-import { Paragraph } from "./paragraph";
-import { Relationships } from "./relationships";
-import { Table } from "./table";
-
-export class HeaderWrapper2 {
-    private readonly header: Header2;
-    private readonly relationships: Relationships;
-
-    constructor(private readonly media: Media) {
-        this.header = new Header2();
-        this.relationships = new Relationships();
-    }
-
-    public addParagraph(paragraph: Paragraph): void {
-        this.header.addParagraph(paragraph);
-    }
-
-    public createParagraph(text?: string): Paragraph {
-        const para = new Paragraph(text);
-        this.addParagraph(para);
-        return para;
-    }
-
-    public addTable(table: Table): void {
-        this.header.addTable(table);
-    }
-
-    public createTable(rows: number, cols: number): Table {
-        return this.header.createTable(rows, cols);
-    }
-
-    public addDrawing(imageData: IMediaData): void {
-        this.header.addDrawing(imageData);
-    }
-
-    public createImage(image: string): void {
-        const mediaData = this.media.addMedia(image, this.relationships.RelationshipCount);
-        this.relationships.createRelationship(
-            mediaData.referenceId,
-            "http://schemas.openxmlformats.org/officeDocument/2006/relationships/image",
-            `media/${mediaData.fileName}`,
-        );
-        this.addDrawing(mediaData);
-    }
-
-    public get Header(): Header2 {
-        return this.header;
-    }
-
-    public get Relationships(): Relationships {
-        return this.relationships;
-    }
-}
