--- src/file/header2/index.ts
@@ -1,1 +1,0 @@
-export * from "./header";
